Revert "change hexToInt to 32-bit"
This reverts commit b406f4d9aa0b9ba5c5bbba229133d603e3b55f25.

+++ chapter_16/Text_Listings/convertHex_C/hexToInt.c
@@ -1,4 +1,4 @@
-// Converts hex character string to int
+// Converts hex character string to long int
 // Returns number of characters
  
 #include "hexToInt.h"
@@ -6,9 +6,9 @@
 #define INTPART 0x0f  // also works for lowercase
 #define NUL '\0'
 
-int hexToInt(int *intPtr, char *stringPtr) {
+int hexToInt(long int *intPtr, char *stringPtr) {
     char current;
-    int result;
+    long int result;
     int count;
 
     count = 0;
